:truck: 중력 로직 GravityComponent 로 이동
