+++ Runelight-Smash-2021/Assets/Scenes/SampleScene.unity
@@ -1442,6 +1442,7 @@ GameObject:
   - component: {fileID: 2025968847}
   - component: {fileID: 2025968848}
   - component: {fileID: 2025968852}
+  - component: {fileID: 2025968853}
   - component: {fileID: 2025968845}
   - component: {fileID: 2025968850}
   - component: {fileID: 2025968849}
@@ -1987,11 +1988,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7f45e4acb7a8cee4eb91fe97d8476438, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gravity: 20
-  maxFallSpeed: 5
-  onLandEvent:
-    m_PersistentCalls:
-      m_Calls: []
   maxWalkSpeed: 5
   walkAccelerationRate: 20
   groundDecelerationRate: 40
@@ -2025,3 +2021,17 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   velocity: {x: 0, y: 0}
+--- !u!114 &2025968853
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2025968840}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 84f9e9612b8d04245a0d869ad9f55c1b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gravity: 20
+  maxFallSpeed: 5
