+++ Runelight-Smash-2021/Assets/Scripts/Component/GravityComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 84f9e9612b8d04245a0d869ad9f55c1b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 250
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
